updated cal task components
